First commit
At the moment, there are no binaries in this repository. I haven't
decided whether to add binaries directly to this repo, or just offer
build scripts.

Either way, as of this commit, I only have binaries compiled for
  Windows x86
  OSX universal
  Linux x64

Once I have x64 for Windows and x86 for Linux, I'll add a download link
for pre-compiled binaries.

If you're intersted in contributing to this project, please feel free
to submit patches.

Signed-off-by: Nicholas Carlson (ncarlson.dev@gmail.com)